05:04 Card,Food,Restaurant apis completed

+++ substring-foodie/src/main/java/com/substring/foodie/entity/OrderItem.java
@@ -24,7 +24,7 @@ public class OrderItem {
     private FoodItem foodItem;
     private int quantity;
 
-    public int getActualPriceOfOrderItem() {
+    public double getActualPriceOfOrderItem() {
         return quantity * foodItem.actualPrice();
     }
